mostly changed to int

+++ resources/shaders/DepthRTTInstanced.vert
@@ -1,12 +1,14 @@
 #version 330 core
 
 // Input vertex data, different for all executions of this shader.
-layout(location = 0) in vec3 vertexPosition_modelspace;
-layout(location = 3) in vec3 model_matrix_transform;
+layout(location = 0) in ivec3 vertexPosition_modelspace;
+layout(location = 3) in ivec3 model_matrix_transform;
 
 // Values that stay constant for the whole mesh.
 uniform mat4 depthMVP;
 
-void main(){
-	gl_Position =  depthMVP * vec4(vertexPosition_modelspace + model_matrix_transform,1);
+void
+main() {
+    gl_Position =
+        depthMVP * vec4(vertexPosition_modelspace + model_matrix_transform, 1);
 }
